clean distinction between types, hooks, and helpers
