+++ packages/x-charts/src/context/CartesianContextProvider.tsx
@@ -13,7 +13,7 @@ import {
   getExtremumX as getLineExtremumX,
   getExtremumY as getLineExtremumY,
 } from '../LineChart/extremums';
-import { getScale } from '../hooks/useScale';
+import { getScale } from '../internals/getScale';
 import { AxisConfig, AxisDefaultized, ScaleName } from '../models/axis';
 import { DrawingContext } from './DrawingProvider';
 import { SeriesContext } from './SeriesContextProvider';
